+++ s1tiling/libs/exits.py
@@ -31,6 +31,9 @@
 This module lists EXIT codes
 """
 
+import logging
+
+OK                  = 0
 TASK_FAILED         = 66
 DOWNLOAD_ERROR      = 67
 OFFLINE_DATA        = 68
@@ -43,3 +46,25 @@ NO_S1_IMAGE         = 74
 MISSING_SRTM        = 75
 MISSING_GEOID       = 76
 MISSING_APP         = 77
+
+logger = logging.getLogger('s1tiling')
+
+class Situation:
+    """
+    Class to help determine the exit value from processing function
+    """
+    def __init__(self, nb_computation_errors, nb_download_failures, nb_download_timeouts):
+        """
+        constructor
+        """
+        logger.debug('Situation: %s computations errors. %s download failures. %s download timeouts',
+                nb_computation_errors, nb_download_failures, nb_download_timeouts)
+        if nb_computation_errors > 0:
+            self.code = TASK_FAILED
+        elif nb_download_failures > nb_download_timeouts:
+            # So far, timeouts are counted as failures as well
+            self.code = DOWNLOAD_ERROR
+        elif nb_download_timeouts > 0:
+            self.code = OFFLINE_DATA
+        else:
+            self.code = OK